chore: update others tab
